Sanitise AvonCap data columns

AvonCap data has lots of columns which are named in a difficult to remember fashion, composed of data items that have enumerated values with no semantics. This makes displaying them difficult and any filtering done on the raw data inscrutable. Depending on the source of the data some different columns may be present due to differences in the NHS and UoB data sets. The redcap database has some options that may be checklists and some that are radio buttons, both of these end up with mysterious names in the data.

Usage

augment_data(x, ...)

Arguments

x

the raw data from load_data()

...

Named arguments passed on to augment_generic

df

a data frame

...

unnamed parameters are a list of functions, named parameters are passed to those functions (if they match formal arguments).

Value

a tracked dataframe with

Details

This function maps the data into a tidy dataframe with consistently named columns, and named factors where appropriate. If not present in the data the ethnicity

files Most of the sanitisation code is held in the zzz-avoncap-mappings.R file.