Categorical scores for continuous variables

Typically used in regression models with non-linear effects over splines

Usage

derive_continuous_categories(df, v, ...)

Arguments

df

the dataframe.

v

the value set. usually precomputed by the augment framework the value set can be explicitly supplied with v = get_value_sets(df)

...

ignored

Value

a dataframe

Details

Age category - UK demographic data ends at 85, and 65 key cut off in 5 year bands, so 10 year bands age categories end at 85 (N.b.) there is a more principled reason here. Boundaries fall approx 0.1, 0.2, 0.4, 0.6, 0.8 quantiles. Could merge first two groups but outcomes are usually different. Covid vaccination cohorts were in 5 year age groups, but vaccination prioirity was in these groups approximately.

Age of eligibility for vaccines: 65+ Age of pneumovax eligibility

CCI - 4 bands as defined in original Charleson paper: ** https://pubmed.ncbi.nlm.nih.gov/3558716/ ** in https://link.springer.com/article/10.1007/s10654-021-00802-z there is rationale given for not using the charleson score as a continuous value.

Alternate CCI - 0,1,2,3+ is also used as a grouping in the original charleson paper

Rockwood score - Completely independent versus dependent frailty levels.

CURB65 categorisation - As per derivation study (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1746657/): 0-1 consider home treatment; 2 consider admit as inpatient; 3-5 admit, consider ICU.