Give a inferred Alpha, Delta or Omicron status based on time alone.

This relies on date period during which we are very confidence that the only variants circulating are of a given type. These are quite conservative estimates based on the frequency of sequenced cases in the bristol area (according to the Sanger centre and to cases identified in the hospital testing)

Usage

derive_genomic_variant(df, v, ...)

Arguments

df

the dataframe.

v

the value set. usually precomputed by the augment framework the value set can be explicitly supplied with v = get_value_sets(df)

...

ignored

Value

a dataframe

Details

Sanger centre data

Pre-alpha before 05 Dec 2020

Alpha between 13 Feb 2021 and 15 May 2021

Delta between 01 Jun 2021 and 07 Nov 2021

Omicron from 07 Feb 2022 to present