Load data and check structure

Loads the AvonCap data from a set of csv files, which may optionally be qualified by site ('BRI' or 'NBT') and database year ('y1', 'y2', 'y3') as part of the file name. This selects the most recent files earlier than the reproduce_at date and detects whether they are in a set of files.

load_data(
  type,
  subtype = NULL,
  reproduce_at = as.Date(getOption("reproduce.at", default = Sys.Date())),
  merge = NA,
  ...
)

Arguments

type

the file category see valid_inputs() for current list in input directory

reproduce_at

the date at which to cut off newer data files

merge

setting to TRUE forces multiple files be merged into a single data frame by losing mismatching columns.

...

passed to cached may specifically want to use `nocache=TRUE``

file

(optional in some circumstances) the file name component see valid_inputs() for current list in input directory

Value

either a list of dataframes or a single merged dataframe

Details

The files are loaded as csv as checked that files have (A) the same columns, (B) the same type (or are empty) (C) have any major parse issues. It then merges the files into a single dataframe, if possible, otherwise it will return the individually loaded files as a list of dataframes.

Examples

load_data("nhs-extract","deltave")
#> using cached item: /home/vp22681/Data/avoncap/cache/cached-969416e3870a1ceaa091ca0c1e17ce42-8e9145a2881200bf0a27d01f5d5e29d9.rda
#> # A tibble: 17,739 × 238
#>    record_…¹ nhs_n…² admis…³ gender age_a…⁴   imd rockw…⁵ sympt…⁶ fever2 pleur…⁷
#>  * <chr>       <dbl> <chr>    <dbl>   <dbl> <dbl>   <dbl>   <dbl>  <dbl>   <dbl>
#>  1 B02976     6.12e9 17/05/…      2    83.4     3       5       0      1       0
#>  2 B02977     4.87e9 17/05/…      2    32.5     2       6      14      0       1
#>  3 B02978     4.72e9 17/05/…      2    77.4     3       5       3      0       0
#>  4 B02980     6.10e9 17/05/…      1    90.6     3       6       0      0       0
#>  5 B02984     4.78e9 17/05/…      2    60.2     1       4       2      0       0
#>  6 B02993     4.58e9 17/05/…      2    65.6     8       6       1      1       0
#>  7 B02994     4.87e9 17/05/…      1    64.1     2       3       0      0       0
#>  8 B03018     4.72e9 18/05/…      1    74.7     4       4       0      0       0
#>  9 B03020     6.28e9 18/05/…      2    65.4     4       4       7      0       0
#> 10 B03024     4.73e9 18/05/…      2    75.7     9       4       2      0       1
#> # … with 17,729 more rows, 228 more variables: cough2 <dbl>, sput_prod <dbl>,
#> #   dyspnoea <dbl>, tachypnoea2 <dbl>, anosmia <dbl>, ageusia <dbl>,
#> #   dysgeusia <dbl>, fever <dbl>, hypothermia <dbl>, chills <dbl>,
#> #   headache <dbl>, malaise <dbl>, wheeze <dbl>, myalgia <dbl>,
#> #   worse_confusion <dbl>, general_det <dbl>, hr <dbl>, systolic_bp <dbl>,
#> #   diastolic_bp <dbl>, temperature <dbl>, rr <dbl>, pulse_ox <dbl>,
#> #   fio2 <dbl>, ox_on_admission <dbl>, news_2_total <dbl>, …