find most recent files of a specific type

find most recent files of a specific type

most_recent_files(
  type = "",
  subtype = NULL,
  reproduce_at = as.Date(getOption("reproduce.at", default = Sys.Date()))
)

Arguments

type

see valid_inputs() for current list of supported types in input directory

subtype

see valid_inputs() for list of supported filenames

reproduce_at

after this date new files are ignored. This enforces a specific version of the data.

Value

a list of the file paths to the most up to date files of the given type relevant to each site and study year

Examples

# devtools::load_all()

avoncap::set_input("~/Data/avoncap")
#> Found data directory: /home/vp22681/Data/avoncap
#> /home/vp22681/Data/avoncap
avoncap::input("nhs-extract")
#> /home/vp22681/Data/avoncap/nhs-extract

avoncap::all_files()
#> # A tibble: 46 × 7
#>    filename     directory      path           date       hospi…¹ study…² filet…³
#>    <chr>        <chr>          <fs::path>     <date>     <chr>     <int> <chr>  
#>  1 central_y1   avoncap-export …entral_y1.csv 2022-03-22 NA            1 csv    
#>  2 central_y2   avoncap-export …entral_y2.csv 2022-03-22 NA            2 csv    
#>  3 central_y1   avoncap-export …entral_y1.csv 2022-06-20 NA            1 csv    
#>  4 central_y2   avoncap-export …entral_y2.csv 2022-06-20 NA            2 csv    
#>  5 central_y1   avoncap-export …entral_y1.csv 2022-07-11 NA            1 csv    
#>  6 central_y2   avoncap-export …entral_y2.csv 2022-07-11 NA            2 csv    
#>  7 central_y1   avoncap-export …entral_y1.csv 2022-09-21 NA            1 csv    
#>  8 central_y2   avoncap-export …entral_y2.csv 2022-09-21 NA            2 csv    
#>  9 uad-controls avoncap-export …-controls.csv 2022-10-07 NA           NA csv    
#> 10 central_y1   avoncap-export …entral_y1.csv 2022-11-09 NA            1 csv    
#> # … with 36 more rows, and abbreviated variable names ¹​hospital, ²​study_year,
#> #   ³​filetype

# exact match on filename column of all_data()
avoncap::most_recent_files("AvonCAPLRTDCentralDa")
#> # A tibble: 0 × 7
#> # … with 7 variables: filename <chr>, directory <chr>, path <fs::path>,
#> #   date <date>, hospital <chr>, study_year <int>, filetype <chr>

# or matches by lower case startWith on directory
avoncap::most_recent_files("nhs-extract","deltave")
#> # A tibble: 6 × 7
#>   filename       directory   path             date       hospi…¹ study…² filet…³
#>   <chr>          <chr>       <fs::path>       <date>     <chr>     <int> <chr>  
#> 1 deltave_BRI_y1 nhs-extract …tave_BRI_y1.csv 2022-11-30 BRI           1 csv    
#> 2 deltave_BRI_y2 nhs-extract …tave_BRI_y2.csv 2022-11-30 BRI           2 csv    
#> 3 deltave_BRI_y3 nhs-extract …tave_BRI_y3.csv 2022-11-30 BRI           3 csv    
#> 4 deltave_NBT_y1 nhs-extract …tave_NBT_y1.csv 2022-11-30 NBT           1 csv    
#> 5 deltave_NBT_y2 nhs-extract …tave_NBT_y2.csv 2022-11-30 NBT           2 csv    
#> 6 deltave_NBT_y3 nhs-extract …tave_NBT_y3.csv 2022-11-30 NBT           3 csv    
#> # … with abbreviated variable names ¹​hospital, ²​study_year, ³​filetype

avoncap::most_recent_files("metadata")
#> # A tibble: 5 × 7
#>   filename     directory path                 date       hospi…¹ study…² filet…³
#>   <chr>        <chr>     <fs::path>           <date>     <chr>     <int> <chr>  
#> 1 central_y1   metadata  …0-24/central_y1.csv 2022-10-24 NA            1 csv    
#> 2 central_y2   metadata  …0-24/central_y2.csv 2022-10-24 NA            2 csv    
#> 3 central_y3   metadata  …0-24/central_y3.csv 2022-10-24 NA            3 csv    
#> 4 uad-controls metadata  …24/uad-controls.csv 2022-10-24 NA           NA csv    
#> 5 pneumococcal metadata  …29/pneumococcal.csv 2022-11-29 NA           NA csv    
#> # … with abbreviated variable names ¹​hospital, ²​study_year, ³​filetype

avoncap::valid_inputs()
#> # A tibble: 19 × 3
#> # Groups:   type [6]
#>    type            subtype      file           
#>    <chr>           <chr>        <chr>          
#>  1 avoncap-export  central      central_y1     
#>  2 avoncap-export  central      central_y2     
#>  3 avoncap-export  uad-controls uad-controls   
#>  4 ethnicity       NA           y1             
#>  5 ethnicity       NA           y2             
#>  6 metadata        central      central_y1     
#>  7 metadata        central      central_y2     
#>  8 metadata        central      central_y3     
#>  9 metadata        uad-controls uad-controls   
#> 10 metadata        pneumococcal pneumococcal   
#> 11 metadata-review NA           metadata-review
#> 12 nhs-extract     deltave      deltave_BRI_y1 
#> 13 nhs-extract     deltave      deltave_BRI_y2 
#> 14 nhs-extract     deltave      deltave_NBT_y1 
#> 15 nhs-extract     deltave      deltave_NBT_y2 
#> 16 nhs-extract     deltave      deltave_BRI_y3 
#> 17 nhs-extract     deltave      deltave_NBT_y3 
#> 18 nhs-extract     pneumococcal pneumococcal   
#> 19 urine-antigens  NA           uad-cases