Sanitise AvonCap data columns

AvonCap data has lots of columns which are named in a difficult to remember fashion, composed of data items that have enumerated values with no semantics. This makes displaying them difficult and any filtering done on the raw data inscrutable. Depending on the source of the data some different columns may be present due to differences in the NHS and UoB data sets. The redcap database has some options that may be checklists and some that are radio buttons, both of these end up with mysterious names in the data.

normalise_data(
  rawData,
  ethn = try(avoncap::load_data("ethnicity", reproduce_at = Sys.Date()), silent = TRUE),
  remove_mapped = TRUE,
  remove_unmapped = TRUE,
  mappings = .mappings,
  messages = c("files: {files}", "{.total} rows", "analysis date: {reproduce_at}",
    "files from: {date}"),
  reproduce_at = as.Date(getOption("reproduce.at", default = Sys.Date())),
  data_source_info = NULL,
  ...
)

Arguments

rawData

the raw data from load_data()

ethn
the supplementary ethnicity files. but default will use the call avoncap::most_recent_files("ethnicity") to find these files and will- can be set to "NULL" if ethnicity files are not available.

remove_mapped
gets rid of original columns for which we have a mapping (leaving the new versions)

remove_unmapped
gets rid of columns for which we do not have a mapping

mappings
a set of mappings (see zzz-avoncap-mappings.R)

messages
a set of dtrackr glue specs that populate the first box fo the flow chart. (can use {files}, {reproduce_at}, {date}, {.total})

reproduce_at
the date at which to reproduce the analysis to (options(reproduce.at = "..."))

data_source_info

if not null a filename, and the function will write out a file with the details of the input files used.

...
passed onto ggrrr::cached(...). e.g. nocache = TRUE can be used to defeat caching.

Value

a tracked dataframe with

Details

This function maps the data into a tidy dataframe with consistently named columns, and named factors where appropriate. If not present in the data the ethnicity files Most of the sanitisation code is held in the zzz-avoncap-mappings.R file.