A valid set of types of file that can be loaded by load_data(...)

A valid set of types of file that can be loaded by load_data(...)

Usage

valid_inputs()

Value

a dataframe of type, subtype

Examples

# devtools::load_all()

avoncap::set_input("~/Data/avoncap")
#> Found data directory: /home/vp22681/Data/avoncap
#> /home/vp22681/Data/avoncap
avoncap::input("nhs-extract")
#> /home/vp22681/Data/avoncap/nhs-extract

avoncap::all_files()
#> # A tibble: 61 × 7
#>    filename     directory     path       date       hospital study_year filetype
#>    <chr>        <chr>         <fs::path> <date>     <chr>         <int> <chr>   
#>  1 central_y1   avoncap-expo… …al_y1.csv 2022-03-22 NA                1 csv     
#>  2 central_y2   avoncap-expo… …al_y2.csv 2022-03-22 NA                2 csv     
#>  3 central_y1   avoncap-expo… …al_y1.csv 2022-06-20 NA                1 csv     
#>  4 central_y2   avoncap-expo… …al_y2.csv 2022-06-20 NA                2 csv     
#>  5 central_y1   avoncap-expo… …al_y1.csv 2022-07-11 NA                1 csv     
#>  6 central_y2   avoncap-expo… …al_y2.csv 2022-07-11 NA                2 csv     
#>  7 central_y1   avoncap-expo… …al_y1.csv 2022-09-21 NA                1 csv     
#>  8 central_y2   avoncap-expo… …al_y2.csv 2022-09-21 NA                2 csv     
#>  9 uad-controls avoncap-expo… …trols.csv 2022-10-07 NA               NA csv     
#> 10 central_y1   avoncap-expo… …al_y1.csv 2022-11-09 NA                1 csv     
#> # ℹ 51 more rows

# exact match on filename column of all_data()
avoncap::most_recent_files("AvonCAPLRTDCentralDa")
#> # A tibble: 0 × 7
#> # ℹ 7 variables: filename <chr>, directory <chr>, path <fs::path>, date <date>,
#> #   hospital <chr>, study_year <int>, filetype <chr>

# or matches by lower case startWith on directory
avoncap::most_recent_files("nhs-extract","deltave")
#> # A tibble: 6 × 7
#>   filename       directory   path        date       hospital study_year filetype
#>   <chr>          <chr>       <fs::path>  <date>     <chr>         <int> <chr>   
#> 1 deltave_BRI_y1 nhs-extract …BRI_y1.csv 2022-12-07 BRI               1 csv     
#> 2 deltave_BRI_y2 nhs-extract …BRI_y2.csv 2022-12-07 BRI               2 csv     
#> 3 deltave_BRI_y3 nhs-extract …BRI_y3.csv 2022-12-07 BRI               3 csv     
#> 4 deltave_NBT_y1 nhs-extract …NBT_y1.csv 2022-12-07 NBT               1 csv     
#> 5 deltave_NBT_y2 nhs-extract …NBT_y2.csv 2022-12-07 NBT               2 csv     
#> 6 deltave_NBT_y3 nhs-extract …NBT_y3.csv 2022-12-07 NBT               3 csv     

avoncap::most_recent_files("metadata")
#> # A tibble: 5 × 7
#>   filename     directory path            date       hospital study_year filetype
#>   <chr>        <chr>     <fs::path>      <date>     <chr>         <int> <chr>   
#> 1 central_y1   metadata  …central_y1.csv 2022-10-24 NA                1 csv     
#> 2 central_y2   metadata  …central_y2.csv 2022-10-24 NA                2 csv     
#> 3 central_y3   metadata  …central_y3.csv 2022-10-24 NA                3 csv     
#> 4 uad-controls metadata  …d-controls.csv 2022-10-24 NA               NA csv     
#> 5 pneumococcal metadata  …eumococcal.csv 2022-11-29 NA               NA csv     

avoncap::valid_inputs()
#> # A tibble: 21 × 3
#> # Groups:   type [6]
#>    type           subtype      file        
#>    <chr>          <chr>        <chr>       
#>  1 avoncap-export central      central_y1  
#>  2 avoncap-export central      central_y2  
#>  3 avoncap-export uad-controls uad-controls
#>  4 avoncap-export central      central_y3  
#>  5 ethnicity      NA           y1          
#>  6 ethnicity      NA           y2          
#>  7 metadata       central      central_y1  
#>  8 metadata       central      central_y2  
#>  9 metadata       central      central_y3  
#> 10 metadata       uad-controls uad-controls
#> # ℹ 11 more rows