Group pneumo serotypes according to e.g. vaccine coverage

A range of useful serotype groups is defined in the list uad_groups. The default_pcv_map gives a set of mappings to group headings that gives the overall serotype distribution by vaccine.

Usage

derive_pcv_groupings(
  df,
  ...,
  pcv_map = uad_pcv_map,
  not_matched = "Other",
  col_name = "pneumo.pcv_group"
)

Arguments

df

the normalised urine antigen data

...

ignored

pcv_map

a 2 column data frame mapping group to uad_analysis

not_matched

what to call the column of non-matched serotypes? Default is Other, but Non vaccine type might be preferred.

col_name

the target column name for the pcv grouping (defaults to pneumo.pcv_group)

Value

an augmented data frame with an additional column defined by col_name

Details

The logic employed in combining elements is:

any(result == "Unknown") ~ "Unknown"

any(result == "Positive") ~ "Positive"

all(result == "Negative") ~ "Negative"

TRUE ~ "Other"